requirements.txt
pyqrcode
pypng
